requirements.txt
pyqrcode
pypng
